增加商店；重写Demo

+++ addons/grid_base_inventory_system/core/view/item/drop_area_view.gd
@@ -19,6 +19,6 @@ func resize() -> void:
 ## 输入控制
 func _gui_input(event: InputEvent) -> void:
 	if event.is_action_pressed(GBIS.input_click):
-		if GBIS.moving_item_service.moving_item:
+		if GBIS.has_moving_item() and GBIS.moving_item_service.moving_item.can_drop():
 			GBIS.moving_item_service.moving_item.drop()
 			GBIS.moving_item_service.clear_moving_item()
